changing field type for contract_year

+++ contracts/models.py
@@ -60,7 +60,7 @@ class Contract(models.Model):
     piid = models.CharField(max_length=128) #index this field
     contract_start = models.DateField(null=True, blank=True)
     contract_end = models.DateField(null=True, blank=True)
-    contract_year = models.DecimalField(max_digits=1, decimal_places=0, null=True, blank=True)
+    contract_year = models.PositiveSmallIntegerField(null=True, blank=True)
     vendor_name = models.CharField(max_length=128)
     labor_category = models.TextField() #index this field
     education_level = models.CharField(choices=EDUCATION_CHOICES, max_length=5, null=True, blank=True)
